pyproject.toml
[project]
name = "qwen-model-loader"
version = "0.1.0"
description = "Offline Qwen GGUF loader with OpenAI/Ollama compatible APIs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "orjson>=3.10.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
legacy = [
  "llama-cpp-python>=0.3.1",
]
test = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
